+++ packages/ripple-binary-codec/src/types/currency.ts
@@ -29,9 +29,7 @@ function isIsoCode(iso: string): boolean {
 function isoCodeFromHex(code: Buffer): string | null {
   const iso = code.toString()
   if (iso === 'XRP') {
-    throw new Error(
-      'Disallowed currency code: to indicate the currency XRP you must use 20 bytes of 0s',
-    )
+    return null
   }
   if (isIsoCode(iso)) {
     return iso