+++ packages/ripple-binary-codec/test/hash.test.js
@@ -50,10 +50,9 @@ describe('Hash256', function () {
 })
 
 describe('Currency', function () {
-  test('Will throw an error for dodgy XRP', function () {
-    expect(() =>
-      Currency.from('0000000000000000000000005852500000000000'),
-    ).toThrow()
+  test('Decoding allows dodgy XRP without throwing', function () {
+    const currencyCode = '0000000000000000000000005852500000000000'
+    expect(Currency.from(currencyCode).toJSON()).toBe(currencyCode)
   })
   test('Currency with lowercase letters decode to hex', () => {
     expect(Currency.from('xRp').toJSON()).toBe(